oracle_spatial_queries/requirements.txt
cx_Oracle
python-dotenv
geopandas
geojson